Special/requirements.txt
pyserial
numpy
scipy